Refactor unit test class names to mirror rule classes
